+++ include/frontend/PIR.h
@@ -92,6 +92,7 @@ namespace panther{
 						EVO_NODISCARD auto operator==(const Param& rhs) const noexcept -> bool;
 					};
 
+
 					std::vector<Param> params;
 					PIR::TypeVoidableID returnType;