+++ src/frontend/Tokenizer.cpp
@@ -168,6 +168,8 @@ namespace panther{
 			else if(ident_name == "def")  { this->create_token(Token::Kind::KeywordDef); }
 			else if(ident_name == "func") { this->create_token(Token::Kind::KeywordFunc); }
 
+			else if(ident_name == "Type") { this->create_token(Token::Kind::KeywordType); }
+
 			else if(ident_name == "return")      { this->create_token(Token::Kind::KeywordReturn); }
 			else if(ident_name == "unreachable") { this->create_token(Token::Kind::KeywordUnreachable); }
 			else if(ident_name == "if")          { this->create_token(Token::Kind::KeywordIf); }